Moving some more logic to SqlRepository, correcting Sqlite queries
